refactor(cli): split parse tests by subcommand
Move the 1.2k-line cli_tests.rs dump into
cli_tests/{serve,download,transcribe,batch,parse}.rs.
Shared env lock and restore helper stay in cli_tests/mod.rs.
