+++ crates/gigastt/src/cli_tests/serve.rs
@@ -1,12 +1,4 @@
-//! CLI parse tests for `gigastt` subcommands. Homed next to `main.rs`.
-
-use super::serve::ServeProfile;
 use super::*;
-use gigastt::server;
-
-// Serialize tests that mutate process env vars to avoid races under
-// cargo test's default multi-threaded harness (used by tarpaulin).
-static ENV_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
 
 #[test]
 fn test_cli_serve_parsing() {
@@ -26,19 +18,6 @@ fn test_cli_serve_parsing() {
             assert_eq!(model_variant, None);
         }
         _ => panic!("expected Serve"),
-    }
-}
-
-// Restore a captured env value when dropped, so an env-mutating test never
-// leaks `GIGASTT_ENCODER_INTRA_THREADS` to a sibling test (clap reads the
-// process environment). Paired with `ENV_LOCK` to serialize these tests.
-struct EnvRestore(&'static str, Option<String>);
-impl Drop for EnvRestore {
-    fn drop(&mut self) {
-        match &self.1 {
-            Some(v) => unsafe { std::env::set_var(self.0, v) },
-            None => unsafe { std::env::remove_var(self.0) },
-        }
     }
 }
 
@@ -127,32 +106,6 @@ fn test_cli_serve_encoder_intra_threads_env() {
 }
 
 #[test]
-fn test_cli_transcribe_encoder_intra_threads_flag() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let _restore = EnvRestore(
-        "GIGASTT_ENCODER_INTRA_THREADS",
-        std::env::var("GIGASTT_ENCODER_INTRA_THREADS").ok(),
-    );
-    unsafe {
-        std::env::remove_var("GIGASTT_ENCODER_INTRA_THREADS");
-    }
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "audio.wav",
-        "--encoder-intra-threads",
-        "3",
-    ]);
-    match cli.command {
-        Commands::Transcribe {
-            encoder_intra_threads,
-            ..
-        } => assert_eq!(encoder_intra_threads, Some(3)),
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
 fn test_cli_serve_model_variant_override() {
     let cli = Cli::parse_from(["gigastt", "serve", "--model-variant", "e2e_rnnt"]);
     match cli.command {
@@ -171,220 +124,6 @@ fn test_cli_serve_model_variant_explicit_rnnt() {
             assert_eq!(model_variant, Some(ModelVariant::Rnnt));
         }
         _ => panic!("expected Serve"),
-    }
-}
-
-#[test]
-fn test_cli_download_parsing() {
-    let cli = Cli::parse_from(["gigastt", "download", "--model-dir", "/tmp/models"]);
-    match cli.command {
-        Commands::Download {
-            model_dir,
-            model_variant,
-            ..
-        } => {
-            assert_eq!(model_dir, "/tmp/models");
-            assert_eq!(model_variant, ModelVariant::Rnnt);
-        }
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_cli_download_model_variant_override() {
-    let cli = Cli::parse_from(["gigastt", "download", "--model-variant", "e2e_rnnt"]);
-    match cli.command {
-        Commands::Download { model_variant, .. } => {
-            assert_eq!(model_variant, ModelVariant::E2eRnnt);
-        }
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_cli_cache_gc_parsing() {
-    let cli = Cli::try_parse_from([
-        "gigastt",
-        "cache-gc",
-        "--model-dir",
-        "/tmp/models",
-        "--dry-run",
-        "--dedupe",
-    ])
-    .expect("parse cache-gc");
-    match cli.command {
-        Commands::CacheGc {
-            model_dir,
-            dry_run,
-            dedupe,
-        } => {
-            assert_eq!(model_dir, "/tmp/models");
-            assert!(dry_run);
-            assert!(dedupe);
-        }
-        _ => panic!("expected CacheGc"),
-    }
-}
-
-#[test]
-fn test_cli_quantize_parsing() {
-    let cli = Cli::parse_from(["gigastt", "quantize", "--force"]);
-    match cli.command {
-        Commands::Quantize { force, .. } => {
-            assert!(force);
-        }
-        _ => panic!("expected Quantize"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_parsing() {
-    let cli = Cli::parse_from(["gigastt", "transcribe", "audio.wav"]);
-    match cli.command {
-        Commands::Transcribe {
-            file,
-            model_variant,
-            format,
-            output,
-            ..
-        } => {
-            assert_eq!(file, "audio.wav");
-            // No --model-variant → None (auto-detect from disk).
-            assert_eq!(model_variant, None);
-            assert_eq!(format, "txt");
-            assert!(output.is_none());
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_format_and_output() {
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "audio.wav",
-        "--format",
-        "srt",
-        "-o",
-        "out.srt",
-    ]);
-    match cli.command {
-        Commands::Transcribe {
-            file,
-            format,
-            output,
-            ..
-        } => {
-            assert_eq!(file, "audio.wav");
-            assert_eq!(format, "srt");
-            assert_eq!(output, Some("out.srt".to_string()));
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_subtitle_options() {
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "audio.wav",
-        "--format",
-        "vtt",
-        "--max-chars-per-line",
-        "60",
-        "--max-words-per-line",
-        "10",
-        "--word-timestamps",
-    ]);
-    match cli.command {
-        Commands::Transcribe {
-            format,
-            max_chars_per_line,
-            max_words_per_line,
-            word_timestamps,
-            ..
-        } => {
-            assert_eq!(format, "vtt");
-            assert_eq!(max_chars_per_line, Some(60));
-            assert_eq!(max_words_per_line, Some(10));
-            assert!(word_timestamps);
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_stereo_speakers_flag() {
-    let cli = Cli::parse_from(["gigastt", "transcribe", "audio.wav", "--stereo-speakers"]);
-    match cli.command {
-        Commands::Transcribe {
-            stereo_speakers, ..
-        } => {
-            assert!(stereo_speakers);
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_stereo_speakers_defaults_off() {
-    let cli = Cli::parse_from(["gigastt", "transcribe", "audio.wav"]);
-    match cli.command {
-        Commands::Transcribe {
-            stereo_speakers, ..
-        } => {
-            assert!(!stereo_speakers);
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_codec_flags() {
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "call.ulaw",
-        "--codec",
-        "pcmu",
-        "--sample-rate",
-        "8000",
-    ]);
-    match cli.command {
-        Commands::Transcribe {
-            codec, sample_rate, ..
-        } => {
-            assert_eq!(codec.as_deref(), Some("pcmu"));
-            assert_eq!(sample_rate, Some(8000));
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_codec_requires_sample_rate() {
-    // clap must reject `--codec` without `--sample-rate` before any engine
-    // work happens.
-    let result = Cli::try_parse_from(["gigastt", "transcribe", "call.ulaw", "--codec", "pcmu"]);
-    assert!(result.is_err(), "--codec without --sample-rate must fail");
-}
-
-#[test]
-fn test_cli_transcribe_sample_rate_alone_is_allowed() {
-    // `--sample-rate` without `--codec` parses (it is simply unused), so
-    // scripts can always append both flags uniformly.
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "audio.wav",
-        "--sample-rate",
-        "8000",
-    ]);
-    match cli.command {
-        Commands::Transcribe { codec, .. } => assert!(codec.is_none()),
-        _ => panic!("expected Transcribe"),
     }
 }
 
@@ -434,31 +173,11 @@ fn test_cli_serve_punctuation_override() {
 }
 
 #[test]
-fn test_cli_transcribe_punctuation_off() {
-    let cli = Cli::parse_from(["gigastt", "transcribe", "a.wav", "--punctuation", "off"]);
-    match cli.command {
-        Commands::Transcribe { punctuation, .. } => {
-            assert_eq!(punctuation, PunctuationMode::Off);
-        }
-        _ => panic!("expected Transcribe"),
-    }
-}
-
-#[test]
 fn test_cli_serve_itn_defaults_auto() {
     let cli = Cli::parse_from(["gigastt", "serve"]);
     match cli.command {
         Commands::Serve(ServeArgs { itn, .. }) => assert_eq!(itn, ItnMode::Auto),
         _ => panic!("expected Serve"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_itn_override() {
-    let cli = Cli::parse_from(["gigastt", "transcribe", "a.wav", "--itn", "on"]);
-    match cli.command {
-        Commands::Transcribe { itn, .. } => assert_eq!(itn, ItnMode::On),
-        _ => panic!("expected Transcribe"),
     }
 }
 
@@ -503,28 +222,6 @@ fn test_cli_serve_hotwords_default_off() {
             assert_eq!(hotwords_boost, None);
         }
         _ => panic!("expected Serve"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_hotwords_flags() {
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "a.wav",
-        "--hotwords-file",
-        "hw.txt",
-    ]);
-    match cli.command {
-        Commands::Transcribe {
-            hotwords_file,
-            hotwords_default,
-            ..
-        } => {
-            assert_eq!(hotwords_file, Some("hw.txt".to_string()));
-            assert!(!hotwords_default);
-        }
-        _ => panic!("expected Transcribe"),
     }
 }
 
@@ -618,16 +315,6 @@ fn test_cli_serve_jobs_defaults_off() {
         }
         _ => panic!("expected Serve"),
     }
-}
-
-#[test]
-fn test_parse_model_variant_valid_and_invalid() {
-    assert_eq!(parse_model_variant("rnnt").unwrap(), ModelVariant::Rnnt);
-    assert_eq!(
-        parse_model_variant("e2e_rnnt").unwrap(),
-        ModelVariant::E2eRnnt
-    );
-    assert!(parse_model_variant("whisper").is_err());
 }
 
 #[test]
@@ -893,167 +580,6 @@ fn test_cli_serve_cors_and_origin_flags() {
 }
 
 #[test]
-fn test_cli_download_parses() {
-    let cli = Cli::try_parse_from(["gigastt", "download"]).expect("parse");
-    match cli.command {
-        Commands::Download { model_variant, .. } => {
-            assert_eq!(model_variant, ModelVariant::Rnnt);
-        }
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_cli_download_rejects_fp32_flag() {
-    let err = match Cli::try_parse_from(["gigastt", "download", "--fp32"]) {
-        Ok(_) => panic!("fp32 flag must be removed from download"),
-        Err(e) => e,
-    };
-    let msg = err.to_string();
-    assert!(
-        msg.contains("unexpected") || msg.contains("fp32") || msg.contains("unknown"),
-        "expected clap to reject --fp32, got: {msg}"
-    );
-}
-
-#[test]
-fn test_cli_download_progress_defaults_to_human() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let _restore = EnvRestore(
-        "GIGASTT_DOWNLOAD_PROGRESS",
-        std::env::var("GIGASTT_DOWNLOAD_PROGRESS").ok(),
-    );
-    unsafe {
-        std::env::remove_var("GIGASTT_DOWNLOAD_PROGRESS");
-    }
-    let cli = Cli::parse_from(["gigastt", "download"]);
-    match cli.command {
-        Commands::Download { progress, .. } => assert_eq!(progress, ProgressMode::Human),
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_cli_download_progress_json_flag() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let _restore = EnvRestore(
-        "GIGASTT_DOWNLOAD_PROGRESS",
-        std::env::var("GIGASTT_DOWNLOAD_PROGRESS").ok(),
-    );
-    unsafe {
-        std::env::remove_var("GIGASTT_DOWNLOAD_PROGRESS");
-    }
-    let cli = Cli::parse_from(["gigastt", "download", "--progress", "json"]);
-    match cli.command {
-        Commands::Download { progress, .. } => assert_eq!(progress, ProgressMode::Json),
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_cli_download_progress_env_var() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let _restore = EnvRestore(
-        "GIGASTT_DOWNLOAD_PROGRESS",
-        std::env::var("GIGASTT_DOWNLOAD_PROGRESS").ok(),
-    );
-    unsafe {
-        std::env::set_var("GIGASTT_DOWNLOAD_PROGRESS", "json");
-    }
-    let cli = Cli::parse_from(["gigastt", "download"]);
-    match cli.command {
-        Commands::Download { progress, .. } => assert_eq!(progress, ProgressMode::Json),
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_parse_progress_mode_value_parser() {
-    assert_eq!(parse_progress_mode("human").unwrap(), ProgressMode::Human);
-    assert_eq!(parse_progress_mode("json").unwrap(), ProgressMode::Json);
-    assert!(parse_progress_mode("xml").is_err());
-}
-
-#[cfg(feature = "ane")]
-#[test]
-fn test_cli_download_ane_flag() {
-    let cli = Cli::parse_from(["gigastt", "download", "--ane"]);
-    match cli.command {
-        Commands::Download { ane, .. } => assert!(ane),
-        _ => panic!("expected Download"),
-    }
-    // Absent by default.
-    let cli = Cli::parse_from(["gigastt", "download"]);
-    match cli.command {
-        Commands::Download { ane, .. } => assert!(!ane),
-        _ => panic!("expected Download"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_vad_and_itn_flags() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let restores: Vec<EnvRestore> = ["GIGASTT_VAD", "GIGASTT_ITN", "GIGASTT_VAD_THRESHOLD"]
-        .iter()
-        .map(|k| {
-            let r = EnvRestore(k, std::env::var(k).ok());
-            unsafe {
-                std::env::remove_var(k);
-            }
-            r
-        })
-        .collect();
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe",
-        "a.wav",
-        "--vad",
-        "--vad-threshold",
-        "0.6",
-        "--itn",
-        "off",
-    ]);
-    match cli.command {
-        Commands::Transcribe {
-            vad,
-            vad_threshold,
-            itn,
-            ..
-        } => {
-            assert!(vad);
-            assert_eq!(vad_threshold, Some(0.6));
-            assert_eq!(itn, ItnMode::Off);
-        }
-        _ => panic!("expected Transcribe"),
-    }
-    drop(restores);
-}
-
-#[test]
-fn test_cli_rejects_unknown_subcommand() {
-    let res = Cli::try_parse_from(["gigastt", "bogus"]);
-    assert!(res.is_err(), "unknown subcommand must be rejected");
-}
-
-#[test]
-fn test_cli_top_level_long_help_points_to_subcommand_engine_flags() {
-    use clap::CommandFactory;
-    let help = Cli::command().render_long_help().to_string();
-    for needle in [
-        "serve --help",
-        "--punctuation",
-        "--itn",
-        "--vad",
-        "--model-variant",
-    ] {
-        assert!(
-            help.contains(needle),
-            "top-level long help must mention `{needle}`:\n{help}"
-        );
-    }
-}
-
-#[test]
 fn test_cli_serve_rejects_bad_punctuation_value() {
     let res = Cli::try_parse_from(["gigastt", "serve", "--punctuation", "sometimes"]);
     assert!(res.is_err(), "invalid punctuation mode must be rejected");
@@ -1063,147 +589,4 @@ fn test_cli_serve_rejects_bad_punctuation_value() {
 fn test_cli_serve_rejects_bad_itn_value() {
     let res = Cli::try_parse_from(["gigastt", "serve", "--itn", "sometimes"]);
     assert!(res.is_err(), "invalid itn mode must be rejected");
-}
-
-#[test]
-fn test_cli_transcribe_batch_defaults() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let restores: Vec<EnvRestore> = ["GIGASTT_FORMAT", "GIGASTT_BATCH_RETRIES"]
-        .iter()
-        .map(|k| {
-            let r = EnvRestore(k, std::env::var(k).ok());
-            unsafe {
-                std::env::remove_var(k);
-            }
-            r
-        })
-        .collect();
-    let cli = Cli::parse_from(["gigastt", "transcribe-batch", "samples/", "out/"]);
-    match cli.command {
-        Commands::TranscribeBatch {
-            input_dir,
-            output_dir,
-            engine,
-            output,
-        } => {
-            assert_eq!(input_dir, "samples/");
-            assert_eq!(output_dir, "out/");
-            assert_eq!(engine.pool_size, 2);
-            assert_eq!(engine.model_variant, None);
-            assert_eq!(output.format, "txt,json");
-            assert_eq!(output.move_to, None);
-            assert!(!output.delete_source);
-            assert_eq!(output.retries, None);
-        }
-        _ => panic!("expected TranscribeBatch"),
-    }
-    drop(restores);
-}
-
-#[test]
-fn test_cli_transcribe_batch_flags() {
-    let cli = Cli::parse_from([
-        "gigastt",
-        "transcribe-batch",
-        "in/",
-        "out/",
-        "--format",
-        "md,srt",
-        "--move-to",
-        "in/done",
-        "--pool-size",
-        "4",
-        "--retries",
-        "1",
-    ]);
-    match cli.command {
-        Commands::TranscribeBatch { engine, output, .. } => {
-            assert_eq!(engine.pool_size, 4);
-            assert_eq!(output.format, "md,srt");
-            assert_eq!(output.move_to, Some("in/done".to_string()));
-            assert_eq!(output.retries, Some(1));
-        }
-        _ => panic!("expected TranscribeBatch"),
-    }
-}
-
-#[test]
-fn test_cli_transcribe_batch_move_to_conflicts_with_delete_source() {
-    let res = Cli::try_parse_from([
-        "gigastt",
-        "transcribe-batch",
-        "in/",
-        "out/",
-        "--move-to",
-        "done/",
-        "--delete-source",
-    ]);
-    assert!(res.is_err(), "--move-to and --delete-source must conflict");
-}
-
-#[test]
-fn test_cli_watch_defaults() {
-    let _guard = ENV_LOCK.lock().unwrap();
-    let restores: Vec<EnvRestore> = [
-        "GIGASTT_WATCH_POLL_INTERVAL_MS",
-        "GIGASTT_WATCH_SETTLE_POLLS",
-        "GIGASTT_FORMAT",
-    ]
-    .iter()
-    .map(|k| {
-        let r = EnvRestore(k, std::env::var(k).ok());
-        unsafe {
-            std::env::remove_var(k);
-        }
-        r
-    })
-    .collect();
-    let cli = Cli::parse_from(["gigastt", "watch", "in/", "out/"]);
-    match cli.command {
-        Commands::Watch {
-            input_dir,
-            output_dir,
-            poll_interval_ms,
-            settle_polls,
-            engine,
-            output,
-        } => {
-            assert_eq!(input_dir, "in/");
-            assert_eq!(output_dir, "out/");
-            assert_eq!(poll_interval_ms, 1000);
-            assert_eq!(settle_polls, 2);
-            assert_eq!(engine.pool_size, 2);
-            assert_eq!(output.format, "txt,json");
-        }
-        _ => panic!("expected Watch"),
-    }
-    drop(restores);
-}
-
-#[test]
-fn test_cli_watch_flags() {
-    let cli = Cli::parse_from([
-        "gigastt",
-        "watch",
-        "in/",
-        "out/",
-        "--poll-interval-ms",
-        "250",
-        "--settle-polls",
-        "4",
-        "--delete-source",
-    ]);
-    match cli.command {
-        Commands::Watch {
-            poll_interval_ms,
-            settle_polls,
-            output,
-            ..
-        } => {
-            assert_eq!(poll_interval_ms, 250);
-            assert_eq!(settle_polls, 4);
-            assert!(output.delete_source);
-        }
-        _ => panic!("expected Watch"),
-    }
 }
